Added Basic Slash Command Functionality

+++ Src/Dto/League/SummonerDto.cs
@@ -5,5 +5,5 @@ public class SummonerDto
     public int profileIconId { get; set; }
     public double revisionDate { get; set; }
     public string puuid { get; set; }
-    public double summonerLevel { get; set; }
+    public int summonerLevel { get; set; }
 }
